requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
elasticsearch==8.15.0
fastmcp>=0.4.1
httpx>=0.27.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
openai>=1.14.0
anthropic>=0.21.0
pydantic>=2.6.0
aiohttp>=3.9.0
lxml>=5.1.0
google-genai>=0.5.0
